# Using string buffer for faster string concatenation