This increases the performance of the script. For more information see https://www.lua.org/pil/11.6.html. 
Would love to see string buffers in the whole script for faster development and testing. :+1: 

Before:
```
time ./update.sh 

real	0m8,036s
user	0m7,084s
sys	0m0,950s
```

After:
```
time ./update.sh 

real	0m4,633s
user	0m4,596s
sys	0m0,038s
```